feat: split root reducer on data and auth reducers
