+++ public/index.html
@@ -9,7 +9,7 @@
     <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
     <link href="https://fonts.googleapis.com/css2?family=Lato:wght@300;400;700&display=swap" rel="stylesheet" />
-    <title>React App</title>
+    <title>DIMS</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
